+++ model/simobject.py
@@ -1,4 +1,9 @@
+import uuid
+
+
 class SeObject:
     """
     A base object for simulation tasks in the smart-energy-di framework
     """
+    def __init__(self):
+        self._uuid = uuid.uuid4()