+++ test/test_basics.py
@@ -1,4 +1,5 @@
 import unittest
+from uuid import UUID
 
 from model.simobject import SeObject
 
@@ -11,6 +12,11 @@ class SimObjectTestCase(unittest.TestCase):
         self.assertEqual(type(self.seObject), SeObject)
         self.assertEqual(str(type(self.seObject)),
                          "<class 'model.simobject.SeObject'>")
+
+    def test_obj_has_valid_uuid(self):
+        uuid_obj = UUID(str(self.seObject._uuid))
+        self.assertEqual(str(uuid_obj),
+                         str(self.seObject._uuid))
 
 
 if __name__ == '__main__':